refactor: organize rules into separate files
