+++ src/sqlfmt/rules/jinja.py
@@ -1,0 +1,200 @@
+from functools import partial
+
+from sqlfmt import actions
+from sqlfmt.rule import Rule
+from sqlfmt.rules.common import group
+from sqlfmt.token import TokenType
+
+JINJA = [
+    Rule(
+        name="jinja_comment",
+        priority=0,
+        pattern=group(r"\{\#.*?\#\}"),
+        action=actions.add_jinja_comment_to_buffer,
+    ),
+    Rule(
+        name="jinja_set_block_start",
+        priority=100,
+        pattern=group(r"\{%-?\s*set\s+[^=]+?-?%\}"),
+        action=actions.handle_jinja_set_block,
+    ),
+    Rule(
+        name="jinja_set_block_end",
+        priority=101,
+        pattern=group(r"\{%-?\s*endset\s*-?%\}"),
+        action=actions.raise_sqlfmt_bracket_error,
+    ),
+    Rule(
+        name="jinja_if_block_start",
+        priority=200,
+        pattern=group(r"\{%-?\s*if.*?-?%\}"),
+        action=partial(
+            actions.handle_jinja_block,
+            start_name="jinja_if_block_start",
+            end_name="jinja_if_block_end",
+            other_names=[
+                "jinja_elif_block_start",
+                "jinja_else_block_start",
+            ],
+        ),
+    ),
+    Rule(
+        name="jinja_elif_block_start",
+        priority=201,
+        pattern=group(r"\{%-?\s*elif\s+\w+.*?-?%\}"),
+        action=actions.raise_sqlfmt_bracket_error,
+    ),
+    Rule(
+        name="jinja_else_block_start",
+        priority=202,
+        pattern=group(r"\{%-?\s*else\s*-?%\}"),
+        action=actions.raise_sqlfmt_bracket_error,
+    ),
+    Rule(
+        name="jinja_if_block_end",
+        priority=203,
+        pattern=group(r"\{%-?\s*endif\s*-?%\}"),
+        action=actions.raise_sqlfmt_bracket_error,
+    ),
+    Rule(
+        name="jinja_for_block_start",
+        priority=210,
+        pattern=group(r"\{%-?\s*for\s+.*?-?%\}"),
+        action=partial(
+            actions.handle_jinja_block,
+            start_name="jinja_for_block_start",
+            end_name="jinja_for_block_end",
+            other_names=[],
+        ),
+    ),
+    Rule(
+        name="jinja_for_block_end",
+        priority=211,
+        pattern=group(r"\{%-?\s*endfor\s*-?%\}"),
+        action=actions.raise_sqlfmt_bracket_error,
+    ),
+    Rule(
+        name="jinja_macro_block_start",
+        priority=220,
+        pattern=group(r"\{%-?\s*macro\s+\w+.*?-?%\}"),
+        action=partial(
+            actions.handle_jinja_block,
+            start_name="jinja_macro_block_start",
+            end_name="jinja_macro_block_end",
+            other_names=[],
+            end_reset_sql_depth=True,
+        ),
+    ),
+    Rule(
+        name="jinja_macro_block_end",
+        priority=221,
+        pattern=group(r"\{%-?\s*endmacro\s*-?%\}"),
+        action=actions.raise_sqlfmt_bracket_error,
+    ),
+    Rule(
+        name="jinja_test_block_start",
+        priority=230,
+        pattern=group(r"\{%-?\s*test\s+\w+.*?-?%\}"),
+        action=partial(
+            actions.handle_jinja_block,
+            start_name="jinja_test_block_start",
+            end_name="jinja_test_block_end",
+            other_names=[],
+            end_reset_sql_depth=True,
+        ),
+    ),
+    Rule(
+        name="jinja_test_block_end",
+        priority=231,
+        pattern=group(r"\{%-?\s*endtest\s*-?%\}"),
+        action=actions.raise_sqlfmt_bracket_error,
+    ),
+    Rule(
+        name="jinja_snapshot_block_start",
+        priority=240,
+        pattern=group(r"\{%-?\s*snapshot\s+\w+.*?-?%\}"),
+        action=partial(
+            actions.handle_jinja_block,
+            start_name="jinja_snapshot_block_start",
+            end_name="jinja_snapshot_block_end",
+            other_names=[],
+            end_reset_sql_depth=True,
+        ),
+    ),
+    Rule(
+        name="jinja_snapshot_block_end",
+        priority=241,
+        pattern=group(r"\{%-?\s*endsnapshot\s*-?%\}"),
+        action=actions.raise_sqlfmt_bracket_error,
+    ),
+    Rule(
+        name="jinja_materialization_block_start",
+        priority=250,
+        pattern=group(r"\{%-?\s*materialization\s+\w+\s*,.*?-?%\}"),
+        action=partial(
+            actions.handle_jinja_block,
+            start_name="jinja_materialization_block_start",
+            end_name="jinja_materialization_block_end",
+            other_names=[],
+            end_reset_sql_depth=True,
+        ),
+    ),
+    Rule(
+        name="jinja_materialization_block_end",
+        priority=251,
+        pattern=group(r"\{%-?\s*endmaterialization\s*-?%\}"),
+        action=actions.raise_sqlfmt_bracket_error,
+    ),
+    Rule(
+        name="jinja_call_block_start",
+        priority=260,
+        pattern=group(r"\{%-?\s*call\s+(noop_)?statement\(.*?\)\s*-?%\}"),
+        action=partial(
+            actions.handle_jinja_block,
+            start_name="jinja_call_block_start",
+            end_name="jinja_call_block_end",
+            other_names=[],
+            end_reset_sql_depth=True,
+        ),
+    ),
+    Rule(
+        name="jinja_call_block_end",
+        priority=261,
+        pattern=group(r"\{%-?\s*endcall\s*-?%\}"),
+        action=actions.raise_sqlfmt_bracket_error,
+    ),
+    Rule(
+        name="jinja_statement_start",
+        priority=500,
+        pattern=group(r"\{%-?"),
+        action=partial(
+            actions.handle_jinja,
+            start_name="jinja_statement_start",
+            end_name="jinja_statement_end",
+            token_type=TokenType.JINJA_STATEMENT,
+        ),
+    ),
+    Rule(
+        name="jinja_expression_start",
+        priority=510,
+        pattern=group(r"\{\{-?"),
+        action=partial(
+            actions.handle_jinja,
+            start_name="jinja_expression_start",
+            end_name="jinja_expression_end",
+            token_type=TokenType.JINJA_EXPRESSION,
+        ),
+    ),
+    Rule(
+        name="jinja_statement_end",
+        priority=600,
+        pattern=group(r"-?%\}"),
+        action=actions.raise_sqlfmt_bracket_error,
+    ),
+    Rule(
+        name="jinja_expression_end",
+        priority=610,
+        pattern=group(r"-?\}\}"),
+        action=actions.raise_sqlfmt_bracket_error,
+    ),
+]
